feat: Update dependency 'marked' to version 8
Upgraded the 'marked' package from version 4.2.5 to version 8.0.1 to make the application compatible with the latest features provided by the package. This caused a breaking change, which led to the need to remove the ids from the H1, H2, H3, H4 heading tags from ArticleDetail component for precise snapshot testing. Also removed '@types/marked' as it is no longer needed due to the updated 'marked' package version.

close #145